refactor: RecommendedArtists section
